Backend/requirements.txt
asgiref==3.3.1
Deprecated==1.2.13
Django==3.1.3
django-cors-headers==3.11.0
django-redis==5.2.0
djangorestframework==3.12.2
Faker==4.14.2
packaging==21.3
Pillow==9.0.1
psycopg2==2.9.3
pyparsing==3.0.7
python-dateutil==2.8.1
python-dotenv==0.19.2
pytz==2020.4
redis==4.1.4
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
validate-docbr==1.8.2
wrapt==1.14.0
